backend/requirements.txt
# app
fastapi==0.55.1
uvicorn==0.11.3
pydantic==1.4
email-validator==1.1.1
python-multipart==0.0.5
pandas==1.2.4
scipy==1.6.3
numpy==1.20.2
yfinance==0.1.59
psycopg2==2.8.6
# db
databases[postgresql]==0.4.2
SQLAlchemy==1.3.16
alembic==1.4.2
# auth
pyjwt==2.0.1
passlib[bcrypt]==1.7.2
# dev
pytest==6.2.2
pytest-asyncio==0.12.0
httpx==0.12.1
asgi-lifespan==1.0.0